requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
PyPDF2==3.0.1
reportlab==4.0.7
pdf2image==1.16.3
Pillow==10.1.0
pdfplumber==0.10.3
supabase==2.4.0
PyMuPDF==1.23.14
python-magic==0.4.27
Flask-Login==0.6.3
bcrypt==4.1.2
python-dateutil==2.8.2
Flask-WTF==1.2.1
WTForms==3.1.1